Refactor JS : export logic into separate files and load it only when necessary.
