+++ templates/base.html.twig
@@ -171,9 +171,10 @@
 {% block body %}
 {% endblock %}
 
-{{ encore_entry_script_tags('app') }}
-
 {% block javascripts %}
 {% endblock %}
+
+{{ encore_entry_script_tags('app') }}
+
 </body>
 </html>
