+++ webpack.config.js
@@ -24,6 +24,9 @@ Encore
      * and one CSS file (e.g. app.css) if your JavaScript imports CSS.
      */
     .addEntry('app', './assets/app.js')
+    .addEntry('imageUpload', './assets/imageUpload.js')
+    .addEntry('loadMore', './assets/loadMore.js')
+    .addEntry('commentsForm', './assets/commentsForm.js')
 
     // enables the Symfony UX Stimulus bridge (used in assets/bootstrap.js)
     // .enableStimulusBridge('./assets/controllers.json')
